Employee Management Project
in this project I used Spring, React and PostgreSQL. i have completed most of the services in the backend part. i will be constantly making updates to the front end part. right now the user can view, delete, update employees, but there is more than that in the spring service and for this I continue my work on the frontend part.